requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-mgmt-core==1.4.0
azure-mgmt-monitor==5.0.1
azure-mgmt-redis==14.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
pywin32==306
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.5.0
urllib3==1.26.15
